64bit support

+++ include/osmium/osm/types.hpp
@@ -37,7 +37,7 @@ enum osm_object_type_t {
 * numbers and still be reasonably space efficient. As the %OSM database is
 * growing rapidly, 64 bit IDs will be needed at some point!
 */
-typedef int32_t  osm_object_id_t;    ///< type for %OSM object (node, way, or relation) IDs
+typedef int64_t  osm_object_id_t;    ///< type for %OSM object (node, way, or relation) IDs
 typedef uint32_t osm_version_t;      ///< type for %OSM object version number
 typedef int32_t  osm_changeset_id_t; ///< type for %OSM changeset IDs
 typedef int32_t  osm_user_id_t;      ///< type for %OSM user IDs
